Fixed md2ie for the latest csString changes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7942 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/impexp/md2/md2ie.cpp
@@ -236,7 +236,7 @@ iModelData *csModelConverterMD2::Load (UByte *Buffer, ULong Size)
   in.SetPosition (Header.SkinOffset);
   for (i = 0; i < Header.SkinCount; i++)
   {
-    csString *s = new csString (SIZEOF_MD2SKINNAME);
+    csString *s = new csString ((unsigned int)SIZEOF_MD2SKINNAME);
     in.Read (s->GetData (), SIZEOF_MD2SKINNAME);
     SkinNames.Push (s);
   }
